fix: set PWM to 50 Hz and fix uint32 overflow in duty calculation
Changed PWM_PERIOD_NS from 25 µs (40 kHz) to 20 ms (50 Hz) for
standard RC/ESC compatibility. Fixed uint32_t overflow in duty cycle
calculation — with 20 ms period, abs(value)*period overflowed ~215,
causing duty to wrap to near-zero every ~215 counts. Cast intermediate
to uint64_t before dividing.

+++ src/motor/tb6612_driver.cpp
@@ -93,12 +93,12 @@ void Tb6612Driver::set_channel(unsigned idx_in1, unsigned idx_in2,
         set_pin(idx_in1, new_dir > 0 ? 1 : 0);
         set_pin(idx_in2, new_dir > 0 ? 0 : 1);
         pwm_write_duty(fd_duty,
-            static_cast<uint32_t>(value > 0 ? value : -value)
-            * Config::Tb6612::PWM_PERIOD_NS / 1000);
+            static_cast<uint32_t>(static_cast<uint64_t>(value > 0 ? value : -value)
+            * Config::Tb6612::PWM_PERIOD_NS / 1000));
     } else {
         pwm_write_duty(fd_duty,
-            static_cast<uint32_t>(value > 0 ? value : -value)
-            * Config::Tb6612::PWM_PERIOD_NS / 1000);
+            static_cast<uint32_t>(static_cast<uint64_t>(value > 0 ? value : -value)
+            * Config::Tb6612::PWM_PERIOD_NS / 1000));
     }
 
     dir = new_dir;
